Refactoring of the general folder structure of the solution
