+++ src/OnionCrafter.Specification/BaseSpecification.cs
@@ -1,10 +1,5 @@
 ﻿using OnionCrafter.Base.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Linq.Expressions;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace OnionCrafter.Specification
 {
