+++ src/OnionCrafter.Specification/Context/IDBContext.cs
@@ -2,13 +2,8 @@
 using Microsoft.EntityFrameworkCore.ChangeTracking;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using OnionCrafter.Base.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace OnionCrafter.Specification.Repository
+namespace OnionCrafter.Specification.Context
 {
     public interface IDBContext : IDisposable
     {
