+++ src/OnionCrafter.Specification/DependencyInjection/ServiceExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using OnionCrafter.Base.Services;
-using OnionCrafter.Specification.Repository;
 using OnionCrafter.Specification.Repository.Cache;
+using OnionCrafter.Specification.UnitOfWork;
 
 namespace OnionCrafter.Specification.DependencyInjection
 {
@@ -9,8 +9,8 @@ namespace OnionCrafter.Specification.DependencyInjection
     {
         public static IServiceCollection AddRepositoryFactory(this IServiceCollection services, Action<UnitOfWorkOptions> unitOfWorkOptions, Action<RepositoryContainerOptions> repositoryContainerOptions)
         {
-            services.AddTypedSingletonWithOptions<RepositoryContainerOptions>(typeof(IRepositoryContainer),typeof(RepositoryContainer), repositoryContainerOptions);
-            services.AddTypedSingletonWithOptions<UnitOfWorkOptions>(typeof(IUnitOfWork<>),typeof(UnitOfWork<>), unitOfWorkOptions );
+            services.AddTypedSingletonWithOptions<RepositoryContainerOptions>(typeof(IRepositoryContainer), typeof(RepositoryContainer), repositoryContainerOptions);
+            services.AddTypedSingletonWithOptions<UnitOfWorkOptions>(typeof(IUnitOfWork<>), typeof(UnitOfWork<>), unitOfWorkOptions);
             return services;
         }
         public static IServiceCollection AddRepositoryFactory(this IServiceCollection services, Action<UnitOfWorkOptions> unitOfWorkOptions)
