+++ src/OnionCrafter.Specification/Repository/IBaseRepository.cs
@@ -1,12 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using OnionCrafter;
-using OnionCrafter.Base.Entities;
-
-namespace OnionCrafter.Specification
+﻿namespace OnionCrafter.Specification.Repository
 {
     public enum RepositoryPrivilegesType
     {
