+++ src/OnionCrafter.Specification/Repository/ICompleteRepository.cs
@@ -1,10 +1,5 @@
 ﻿using OnionCrafter.Base.Entities;
 using OnionCrafter.Specification.Repository;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace OnionCrafter.Specification
 {
