+++ src/OnionCrafter.Specification/Repository/IReadRepository.cs
@@ -1,4 +1,5 @@
 ﻿using OnionCrafter.Base.Entities;
+using OnionCrafter.Specification.Repository.Base;
 
 namespace OnionCrafter.Specification.Repository
 {
