+++ src/OnionCrafter.Specification/Repository/IWriteRepository.cs
@@ -1,9 +1,4 @@
 ﻿using OnionCrafter.Base.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace OnionCrafter.Specification.Repository
 {
