+++ src/OnionCrafter.Specification/Repository/Repository.cs
@@ -1,9 +1,5 @@
 ﻿using OnionCrafter.Base.Entities;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+using OnionCrafter.Specification.Context;
 
 namespace OnionCrafter.Specification.Repository
 {
