+++ src/OnionCrafter.Specification/UnitOfWork/UnitOfWorkOptions.cs
@@ -1,11 +1,6 @@
 ﻿using OnionCrafter.Base.Services;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace OnionCrafter.Specification.Repository
+namespace OnionCrafter.Specification.UnitOfWork
 {
     //IServiceOptions debe de tener tipado
     public class UnitOfWorkOptions : IServiceOptions
