+++ src/OnionCrafter.Specification/Utils/LoggingUtils.cs
@@ -1,9 +1,4 @@
 ﻿using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace OnionCrafter.Specification.Utils
 {
